Moved all test helper into single package
