+++ controllers/registration_test.go
@@ -5,8 +5,7 @@ import (
 	"net/url"
 	"strings"
 
-	. "google-scraper/tests/fabricators"
-	. "google-scraper/tests/testing_helpers"
+	. "google-scraper/tests"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
