+++ forms/session_form_test.go
@@ -3,8 +3,7 @@ package forms_test
 import (
 	"google-scraper/forms"
 	_ "google-scraper/initializers"
-	. "google-scraper/tests/fabricators"
-	. "google-scraper/tests/testing_helpers"
+	. "google-scraper/tests"
 
 	"github.com/beego/beego/v2/core/validation"
 	. "github.com/onsi/ginkgo"
