+++ helpers/password_test.go
@@ -2,8 +2,7 @@ package helpers_test
 
 import (
 	"google-scraper/helpers"
-	. "google-scraper/tests/fabricators"
-	. "google-scraper/tests/testing_helpers"
+	. "google-scraper/tests"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
