+++ models/user_test.go
@@ -3,8 +3,7 @@ package models_test
 import (
 	_ "google-scraper/initializers"
 	"google-scraper/models"
-	. "google-scraper/tests/fabricators"
-	. "google-scraper/tests/testing_helpers"
+	. "google-scraper/tests"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
