--- tests/fabricators/keyword_fabricator.go
@@ -1,20 +1,0 @@
-package fabricators
-
-import (
-	. "google-scraper/models"
-
-	"github.com/beego/beego/v2/client/orm"
-	"github.com/beego/beego/v2/core/logs"
-)
-
-func FabricateKeyword(name string, user *User) Keyword {
-	o := orm.NewOrm()
-	keyword := Keyword{Name: name, User: user}
-
-	_, err := o.Insert(&keyword)
-	if err != nil {
-		logs.Error("Creating keyword failed: ", err)
-	}
-
-	return keyword
-}
